requirements.txt
argon2==0.1.10
argon2-cffi==18.3.0
cffi==1.11.5
Django==2.1.5
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.7
django-multiselectfield==0.1.8
django-widget-tweaks==1.4.3
djangorestframework==3.9.1
mysqlclient==1.4.1
Pillow==5.4.1
pycountry==18.12.8
pycparser==2.19
PyMySQL==0.9.3
pytz==2018.9
reportlab==3.5.12
six==1.12.0
